fastapi/requirements.txt
# Web Framework
fastapi

# Server
uvicorn

# Environment
python-dotenv

# Data Models
pydantic
pillow

# Lint
black

# Youtube Search
youtube_search
minify_html

# Language Community Resources
langchain-community
unstructured

# Utilities
plantuml
youtube-transcript-api
starlette
matplotlib
seaborn

# Others
redberry
